requirements/dev.txt
-r web.txt

q
pytest-django==2.9.1
